+++ tests/hazmat/primitives/test_rsa.py
@@ -759,12 +759,15 @@ class TestRSAVerification(object):
         public_key = rsa.RSAPublicNumbers(
             e=public["public_exponent"], n=public["modulus"]
         ).public_key(backend)
-        public_key.verify(
-            binascii.unhexlify(example["signature"]),
-            binascii.unhexlify(example["message"]),
-            padding.PKCS1v15(),
-            hashes.SHA1(),
-        )
+        signature = binascii.unhexlify(example["signature"])
+        message = binascii.unhexlify(example["message"])
+        public_key.verify(signature, message, padding.PKCS1v15(), hashes.SHA1())
+
+        digest = hashes.Hash(hashes.SHA1())
+        digest.update(message)
+        msg_digest = digest.finalize()
+        rec_msg_digest = public_key.recover(signature, padding.PKCS1v15(), hashes.SHA1())
+        assert msg_digest == rec_msg_digest
 
     @pytest.mark.supported(
         only_if=lambda backend: backend.rsa_padding_supported(
@@ -781,6 +784,17 @@ class TestRSAVerification(object):
         with pytest.raises(InvalidSignature):
             public_key.verify(
                 signature, b"incorrect data", padding.PKCS1v15(), hashes.SHA1()
+            )
+
+    def test_invalid_pkcs1v15_signature_recover_wrong_hash_alg(self, backend):
+        private_key = RSA_KEY_512.private_key(backend)
+        public_key = private_key.public_key()
+        signature = private_key.sign(
+            b"sign me", padding.PKCS1v15(), hashes.SHA1()
+        )
+        with pytest.raises(InvalidSignature):
+            public_key.recover(
+                signature, padding.PKCS1v15(), hashes.SHA256()
             )
 
     def test_invalid_signature_sequence_removed(self, backend):
@@ -969,6 +983,17 @@ class TestRSAVerification(object):
                 ),
                 hashes.SHA1(),
             )
+
+    def test_invalid_pss_signature_recover(self, backend):
+        private_key = RSA_KEY_1024.private_key(backend)
+        public_key = private_key.public_key()
+        pss_padding = padding.PSS(
+            mgf=padding.MGF1(algorithm=hashes.SHA1()),
+            salt_length=padding.PSS.MAX_LENGTH,
+        )
+        signature = private_key.sign(b"sign me", pss_padding, hashes.SHA1())
+        with raises_unsupported_algorithm(_Reasons.UNSUPPORTED_PADDING):
+            public_key.recover(signature, pss_padding, hashes.SHA1())
 
     @pytest.mark.supported(
         only_if=lambda backend: backend.rsa_padding_supported(